Move mock listeing module to testing package in root
